Add files via upload
The purpose of this program is to provide a reference on how to script the programming of PIC targets using the MPLAB IPE utility v4.1. This script can aid in situations where automated programming of firmware is desired (e.g. MFG).  This script can be altered for use with other Microchip programmers (PICKit3, Real ICE, etc.). This script can likely be used with previous or future versions of MPLAB IPE but v4.1 was used to develop and test this script.